+++ packages/cms/lib/modules/choices-guide-result-widgets/index.js
@@ -26,6 +26,13 @@ module.exports = {
 			required: true,
 		},
 		{
+			name: 'startWithAllQuestionsAnswered',
+      type: 'boolean',
+			label: 'Begin met alle vragen beantwoord op 50%',
+			required: true,
+      def: false,
+		},
+		{
 			type: 'select',
 			name: 'choicesType',
 			label: 'Weergave van de voorkeuren',
@@ -33,15 +40,17 @@ module.exports = {
 				{
 					label: 'Standaard',
 					value: 'default',
+          showFields: ['choicesPreferenceTitle','choicesNoPreferenceYetTitle', 'choicesWithPercentage','choicesMinLabel','choicesMaxLabel'],
 				},
 				{
 					label: 'Van min naar plus 100',
 					value: 'minus-to-plus-100',
-          showFields: ['choicesPreferenceMinColor', 'choicesPreferenceMaxColor']
+          showFields: ['choicesPreferenceMinColor', 'choicesPreferenceMaxColor','choicesPreferenceTitle','choicesNoPreferenceYetTitle', 'choicesWithPercentage','choicesMinLabel','choicesMaxLabel'],
 				},
 				{
 					label: 'In een vlak',
-					value: 'plane'
+					value: 'plane',
+          showFields: ['choicesPreferenceTitle','choicesNoPreferenceYetTitle', 'choicesWithPercentage','choicesMinLabel','choicesMaxLabel'],
 				}
 			]
 		},
@@ -51,27 +60,70 @@ module.exports = {
       label:    'Kleur van de balken, minimaal',
       help:     'Dit moet (nu nog) in het formaat #123456',
       def:      '#ff9100',
-    },
-    {
+    }, {
       type:     'string',
       name:     'choicesPreferenceMaxColor',
       label:    'Kleur van de balken, maximaal',
       help:     'Dit moet (nu nog) in het formaat #123456',
       def:      '#bed200',
-    },
-    {
+    }, {
       type:     'string',
       name:     'choicesPreferenceTitle',
       label:    'Titel boven de keuzes, met voorkeur',
       help:     'Bijvoorbeeld "Jouw voorkeur is {preferredChoice}"',
       def:      'Jouw voorkeur is {preferredChoice}',
-    },
-    {
+    }, {
       type:     'string',
       name:     'choicesNoPreferenceYetTitle',
       label:    'Titel boven de keuzes, nog geen voorkeur',
       help:     'Bijvoorbeeld "Je hebt nog geen keuze gemaakt"',
       def:      'Je hebt nog geen keuze gemaakt',
+    }, {
+      type:     'string',
+      name:     'choicesMinLabel',
+      label:    'Tekst links bij de balken',
+    }, {
+      type:     'string',
+      name:     'choicesMaxLabel',
+      label:    'Tekst rechts bij de balken',
+    },{
+      type:     'string',
+      name:     'requireLoginTitle',
+      label:    'Titel',
+    }, {
+      type:     'string',
+      name:     'requireLoginDescription',
+      label:    'Beschrijving',
+    }, {
+      type:     'string',
+      name:     'requireLoginButtonTextLogin',
+      label:    'Tekst op de button - je moet nog inloggen',
+    }, {
+      type:     'string',
+      name:     'requireLoginButtonTextLoggedIn',
+      label:    'Tekst op de button - je bent ingelogd',
+    }, {
+      type:     'string',
+      name:     'requireLoginButtonTextAlreadySubmitted',
+      label:    'Tekst op de button - je hebt al gestemd',
+    }, {
+      type:     'string',
+      name:     'requireLoginChangeLoginLinkText',
+      label:    'Tekst voor link \'Opnieuw inloggen\'',
+    }, {
+      type:     'string',
+      name:     'requireLoginLoggedInMessage',
+      label:    '\'Login gelukt\' bericht',
+    }, {
+      type:     'string',
+      name:     'requireLoginAlreadySubmittedMessage',
+      label:    '\'Je hebt al gestemd\' bericht',
+    }, {
+      type: 'boolean',
+      name:     'choicesWithPercentage',
+      label:    'Toon percentage achter de balk',
+      def:      false,
+      
     },
 		{
 			type: 'select',
@@ -91,7 +143,7 @@ module.exports = {
 					label: 'Een formulier met extra gegevens',
 					value: 'form',
 					showTab: [
-						'Form'
+						'form', "requireLogin"
 					]
 				}
 			]
@@ -135,14 +187,12 @@ module.exports = {
       type:  'string',
       name:  'formTitle',
       label: 'Title',
-    },
-    {
+    }, {
       type:     'string',
       name:     'formIntro',
       label:    'Intro',
       textarea: true
-    },
-    {
+    }, {
       name:       'formFields',
       label:      'Form fields',
       type:       'array',
@@ -248,12 +298,17 @@ module.exports = {
       {
         name: 'general',
         label: 'Algemeen',
-        fields: ['choicesGuideId', 'questionGroupId', 'choicesType', 'choicesPreferenceMinColor', 'choicesPreferenceMaxColor', 'choicesPreferenceTitle', 'choicesNoPreferenceYetTitle', 'moreInfoUrl', 'moreInfoLabel', 'submissionType', ]
+        fields: ['choicesGuideId', 'questionGroupId', 'choicesType', 'choicesPreferenceMinColor', 'choicesPreferenceMaxColor', 'choicesPreferenceTitle', 'choicesNoPreferenceYetTitle', 'choicesMinLabel', 'choicesMaxLabel', 'choicesWithPercentage', 'startWithAllQuestionsAnswered', 'moreInfoUrl', 'moreInfoLabel', 'submissionType', ]
       },
       {
         name: 'form',
         label: 'Formulier',
         fields: ['formTitle', 'formIntro', 'formFields', 'beforeUrl', 'beforeLabel', 'afterUrl', 'afterLabel',]
+      },
+      {
+        name: 'reuiredLogin',
+        label: 'Login',
+        fields: ['requireLoginTitle', 'requireLoginDescription', 'requireLoginButtonTextLogin', 'requireLoginButtonTextLoggedIn', 'requireLoginButtonTextAlreadySubmitted', 'requireLoginChangeLoginLinkText', 'requireLoginLoggedInMessage', 'requireLoginAlreadySubmittedMessage',]
       },
 
       /*  {
@@ -273,7 +328,19 @@ module.exports = {
 		self.load = function(req, widgets, next) {
 
 			widgets.forEach((widget) => {
-			  widget.config = JSON.stringify({
+
+        let requireLoginSettings;
+        requireLoginSettings = {};
+        if (widget.requireLoginTitle) requireLoginSettings.title = widget.requireLoginTitle;
+        if (widget.requireLoginDescription) requireLoginSettings.description = widget.requireLoginDescription;
+        if (widget.requireLoginButtonTextLogin) requireLoginSettings.buttonTextLogin = widget.requireLoginButtonTextLogin;
+        if (widget.requireLoginButtonTextLoggedIn) requireLoginSettings.buttonTextLoggedIn = widget.requireLoginButtonTextLoggedIn;
+        if (widget.requireLoginButtonTextAlreadySubmitted) requireLoginSettings.buttonTextAlreadySubmitted = widget.requireLoginButtonTextAlreadySubmitted;
+        if (widget.requireLoginChangeLoginLinkText) requireLoginSettings.changeLoginLinkText = widget.requireLoginChangeLoginLinkText;
+        if (widget.requireLoginLoggedInMessage) requireLoginSettings.loggedInMessage = widget.requireLoginLoggedInMessage;
+        if (widget.requireLoginAlreadySubmittedMessage) requireLoginSettings.alreadySubmittedMessage = widget.requireLoginAlreadySubmittedMessage;
+
+        widget.config = JSON.stringify({
           // req.data.isAdmin
           divId: 'choices-guide-result',
           siteId: req.data.global.siteId,
@@ -281,6 +348,10 @@ module.exports = {
             url: self.apos.settings.getOption(req, 'apiUrl'),
             headers: req.session.jwt ? { 'X-Authorization': 'Bearer ' + req.session.jwt } : {},
             isUserLoggedIn: req.data.loggedIn,
+          },
+          user: {
+            role:  req.data.openstadUser && req.data.openstadUser.role,
+            fullName:  req.data.openstadUser && (req.data.openstadUser.fullName || req.data.openstadUser.firstName + ' ' + req.data.openstadUser.lastName)
           },
           choicesGuideId: widget.choicesGuideId,
           questionGroupId: widget.questionGroupId,
@@ -291,6 +362,10 @@ module.exports = {
               noPreferenceYet: widget.choicesNoPreferenceYetTitle,
             },
             barColor: { min: widget.choicesPreferenceMinColor || null, max: widget.choicesPreferenceMaxColor || null },
+            startWithAllQuestionsAnswered: widget.startWithAllQuestionsAnswered,
+            minLabel: widget.choicesMinLabel,
+            maxLabel: widget.choicesMaxLabel,
+            withPercentage: widget.choicesWithPercentage,
           },
           moreInfoUrl: widget.moreInfoUrl,
           moreInfoLabel: widget.moreInfoLabel,
@@ -305,9 +380,12 @@ module.exports = {
               intro: widget.formIntro,
               fields: widget.formFields,
             },
+            requireLogin: widget.requireLogin,
+            requireLoginSettings,
           },
           preferenceTitle: widget.preferenceTitle,
         });
+        
         widget.openstadComponentsUrl = openstadComponentsUrl;
         const containerId = widget._id;
         widget.containerId = containerId;