+++ packages/cms/lib/modules/resource-form-widgets/public/js/main.js
@@ -31,8 +31,8 @@ if (fieldsetElement) {
     allowBrowse: true,
     files: uploadedFiles,
     server: {
-      process: '/image',
-      fetch: '/fetch-image?img=',
+      process: window.siteUrl + '/image',
+      fetch: window.siteUrl + '/fetch-image?img=',
       revert: null
     },
     labelIdle: "Sleep afbeelding(en) naar deze plek of <span class='filepond--label-action'>klik hier</span>",
@@ -62,6 +62,8 @@ if (fieldsetElement) {
     labelButtonRetryItemProcessing: "Retry",
     labelButtonProcessItem: "Upload"
   }
+
+  console.log('filePond', filePondSettings)
 
   var pond = FilePond.create(fieldsetElement, filePondSettings);
 
@@ -192,7 +194,7 @@ $(document).ready(function () {
               var redirect = $(form).find('.form-redirect-uri').val();
               redirect = redirect.replace(':id', response.id);
               //use href to simulate a link click! Not replace, that doesn't allow for back button to work
-              window.location.href = redirect;
+              window.location.href = window.siteUrl + redirect;
           },
           error:function(response) {
               // "this" the object you passed